Add files via upload
This is a project that will focus on retrieving tweets from both houses of congress (and the President?) and examine three things: 1) how the "elites" / "leaders" of the respective parties talk to their audiences, 2) if Republicans are focused on an ideology while Democrats are spread across many areas of interest, and 3) how the elites and their followers interact with each other on social media. 

While initially focusing on twitter, this project may include or head into other platforms.